% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_generate}
\alias{ts_generate}
\title{Generate ULIDs from timestamps}
\usage{
ts_generate(tsv)
}
\arguments{
\item{tsv}{vector of \code{POSIXct} values}
}
\description{
This function generates a new \href{https://github.com/ulid/spec}{Universally Unique Lexicographically Sortable Identifier} from a vector of
\code{POSIXct} timestamps.
}
\examples{
ts_generate(as.POSIXct("2017-11-01 15:00:00", origin="1970-01-01"))
}
