% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ULIDgenerate}
\alias{ULIDgenerate}
\title{Generate ULID}
\usage{
ULIDgenerate(n = 1L)
}
\arguments{
\item{n}{number of id's to generate (default = \code{1})}
}
\description{
\code{\link[=ULIDgenerate]{ULIDgenerate()}} generates a new []Universally Unique Lexicographically
Sortable Identifier](https://github.com/ulid/spec).
}
\examples{
ULIDgenerate()
}
