% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unmarshal}
\alias{unmarshal}
\title{Unmarshal a ULID into a data frame with timestamp and random bitstring columns}
\usage{
unmarshal(ulids)
}
\arguments{
\item{ulids}{character ULIDs (e.g. created with \code{\link[=ULIDgenerate]{ULIDgenerate()}})}
}
\description{
Unmarshal a ULID into a data frame with timestamp and random bitstring columns
}
\examples{
unmarshal(ULIDgenerate())
}
