% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulid-package.R
\docType{package}
\name{ulid}
\alias{ulid}
\alias{ulid-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/ulid}
\item BugReports: \url{https://gitlab.com/hrbrmstr/ulid/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
