% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/gen.R
\name{ULIDgenerate}
\alias{ULIDgenerate}
\alias{generate}
\alias{ulid_generate}
\title{Generate ULID}
\usage{
ULIDgenerate(n = 1L)

generate(n = 1L)

ulid_generate(n = 1L)
}
\arguments{
\item{n}{number of id's to generate (default = \code{1})}
}
\description{
\code{\link[=ULIDgenerate]{ULIDgenerate()}} generates a new \href{https://github.com/ulid/spec}{Universally Unique Lexicographically Sortable Identifier}.
}
\examples{
ULIDgenerate()
}
